# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Farming API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /predict/soil-moisture:
    post:
      operationId: predictSoilMoistureTree
      tags: [sensor]
      summary: Predict next-step soil moisture (Decision Tree/Random Forest)
      parameters:
        - name: model_type
          in: query
          required: false
          schema:
            type: string
            enum: [decision_tree, random_forest]
            default: random_forest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                temperature: { type: number }
                humidity: { type: number }
                rain: { type: integer }
                prev_moisture: { type: number }
              required: [temperature, humidity, rain, prev_moisture]
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                type: object
                properties:
                  predictedMoisture: { type: number }
                  status: { type: string }
                required: [predictedMoisture, status]
  /predict/forecast:
    post:
      operationId: predictForecast
      tags: [sensor]
      summary: Forecast next value in sequence (Time Series)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                lags:
                  type: array
                  items: { type: number }
                  minItems: 3
                  maxItems: 3
              required: [lags]
      responses:
        "200":
          description: Forecast result
          content:
            application/json:
              schema:
                type: object
                properties:
                  forecast: { type: number }
                  status: { type: string }
                required: [forecast, status]
...